+++ native/ggml-hnx/ggml-hnx-shim.h
@@ -34,6 +34,9 @@ void hnx_ggml_to_float_iq0_75(const void *x, float *y, int64_t k);
 void hnx_ggml_to_float_iq0_5 (const void *x, float *y, int64_t k);
 void hnx_ggml_to_float_iq0_25(const void *x, float *y, int64_t k);
 void hnx_ggml_to_float_int1  (const void *x, float *y, int64_t k);
+void hnx_ggml_to_float_hnx1375(const void *x, float *y, int64_t k);
+void hnx_ggml_to_float_int4  (const void *x, float *y, int64_t k);
+void hnx_ggml_to_float_fp2   (const void *x, float *y, int64_t k);
 
 void hnx_ggml_vec_dot_iq0_9 (int n, float *s, size_t bs, const void *x,
                              size_t bx, const void *y, size_t by, int nrc);
@@ -45,6 +48,12 @@ void hnx_ggml_vec_dot_iq0_25(int n, float *s, size_t bs, const void *x,
                              size_t bx, const void *y, size_t by, int nrc);
 void hnx_ggml_vec_dot_int1  (int n, float *s, size_t bs, const void *x,
                              size_t bx, const void *y, size_t by, int nrc);
+void hnx_ggml_vec_dot_hnx1375(int n, float *s, size_t bs, const void *x,
+                               size_t bx, const void *y, size_t by, int nrc);
+void hnx_ggml_vec_dot_int4(int n, float *s, size_t bs, const void *x,
+                            size_t bx, const void *y, size_t by, int nrc);
+void hnx_ggml_vec_dot_fp2(int n, float *s, size_t bs, const void *x,
+                           size_t bx, const void *y, size_t by, int nrc);
 
 #ifdef __cplusplus
 }